requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
annoy==1.17.1
anyio==3.6.2
asgiref==3.7.2
astunparse==1.6.3
async-timeout==4.0.2
asyncer==0.0.2
attrs==23.1.0

billiard==3.6.4.0
cachetools==5.3.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cmake==3.26.3
coloredlogs==15.0.1
colorgram.py==1.2.0
colormath==3.0.0
colorthief==0.2.1
contourpy==1.0.7
coreapi==2.3.3
coreschema==0.0.4
cycler==0.11.0
DeepImageSearch==1.4
Django==4.2.1
djangorestframework==3.14.0
drf-yasg==1.21.5
fastapi==0.95.2
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.39.4
frozenlist==1.3.3
gast==0.4.0
google-auth==2.19.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.2
gunicorn==20.1.0
h11==0.14.0
h5py==3.8.0
humanfriendly==10.0
idna==3.4
ImageHash==4.3.1
imageio==2.30.0
importlib-metadata==6.6.0
importlib-resources==5.12.0
inflection==0.5.1
itypes==1.2.0
jax==0.4.11
Jinja2==3.1.2
keras==2.12.0
kiwisolver==1.4.4
kombu==5.2.4
lazy-loader==0.2
libclang==16.0.0
llvmlite==0.40.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
ml-dtypes==0.1.0
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numba==0.57.0
numpy==1.24.3
oauthlib==3.2.2
onnxruntime==1.15.0
opencv-python-headless==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
platformdirs==3.5.1
pooch==1.7.0
prompt-toolkit==3.0.38
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.8
PyMatting==1.1.8
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyWavelets==1.4.1
redis==4.5.5
rembg==2.0.37
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
scikit-image==0.20.0
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.27.0
sympy==1.12
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
tifffile==2023.4.12
tqdm==4.65.0
typing-extensions==4.6.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
uvicorn==0.22.0
vine==5.0.0
watchdog==3.0.0
wcwidth==0.2.6
Werkzeug==2.3.4
wrapt==1.14.1
yarl==1.9.2
zipp==3.15.0
